refactor: databasecleanup 패키지 이동 및 기본 설정
