+++ src/main/resources/application.yaml
@@ -1,12 +1,12 @@
 spring:
-    profiles:
-        include: release
     jpa:
         properties:
             hibernate:
                 show_sql: true
                 format_sql: true
                 use_sql_comments: true
+    config:
+        import: security/aws-default.yml
 logging:
     level:
         org:
